+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/doc.go
@@ -1,4 +1,4 @@
-package remotewriteagentconfigsecret
+package remotewriteapiendpointsecret
 
 // This packages creates the remote write agent secret config
 // to be used as values by the prometheus agent app to configure the remote write feature.
