+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/resource_test.go
@@ -1,4 +1,4 @@
-package remotewriteagentconfigsecret
+package remotewriteapiendpointsecret
 
 import (
 	"context"
