+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/test/case-3-kvmconfig.golden
