+++ service/controller/resource/monitoring/remotewriteapiendpointconfigsecret/test/case-5-cluster-api-v1alpha3.golden
