+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/doc.go
@@ -1,4 +1,4 @@
-package remotewriteagentsecret
+package remotewriteapiendpointsecret
 
 // This packages creates the remote write secret
 // to be used by the remote write agent so it can access
