+++ service/controller/resource/monitoring/remotewriteapiendpointsecret/test/case-1-awsconfig.golden
